requirements.txt
appdirs==1.4.3
attrs==19.3.0
beautifulsoup4==4.8.1
black==19.3b0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
html2text==2019.9.26
idna==2.8
lxml==4.4.1
matrix-client==0.3.2
pycparser==2.19
pyOpenSSL==19.0.0
python-dotenv==0.10.3
requests==2.22.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.4
toml==0.10.0
typing==3.7.4.1
urllib3==1.25.6
zulip==0.6.1
zulip-bots==0.6.1
zulipbot==0.0.1
